+++ src/hyperform.js
@@ -21,31 +21,17 @@ import sprintf from './tools/sprintf';
 import version from './version';
 
 
-/* deprecate the old snake_case names
- * TODO: delme before next non-patch release
- */
-function w(name) {
-  const deprecated_message = 'Please use camelCase method names! The name "%s" is deprecated and will be removed in the next non-patch release.';
-  /* global console */
-  console.log(sprintf(deprecated_message, name));
-}
-
-
 /**
  * public hyperform interface:
  */
 function hyperform(form, {
                      classes,
                      debug=false,
-                     extend_fieldset,
                      extendFieldset,
-                     novalidate_on_elements,
                      novalidateOnElements,
-                     prevent_implicit_submit,
-                     preventImplicitSubmit/* TODO: uncomment =false */,
+                     preventImplicitSubmit=false,
                      revalidate,
                      strict=false,
-                     valid_event,
                      validEvent,
                      validateNameless=false,
                    }={}) {
@@ -55,40 +41,20 @@ function hyperform(form, {
   }
   // TODO: clean up before next non-patch release
   if (extendFieldset === undefined) {
-    if (extend_fieldset === undefined) {
-      extendFieldset = ! strict;
-    } else {
-      w('extend_fieldset');
-      extendFieldset = extend_fieldset;
-    }
+    extendFieldset = ! strict;
   }
   if (novalidateOnElements === undefined) {
-    if (novalidate_on_elements === undefined) {
-      novalidateOnElements = ! strict;
-    } else {
-      w('novalidate_on_elements');
-      novalidateOnElements = novalidate_on_elements;
-    }
+    novalidateOnElements = ! strict;
   }
   if (preventImplicitSubmit === undefined) {
-    if (prevent_implicit_submit === undefined) {
-      preventImplicitSubmit = false;
-    } else {
-      w('prevent_implicit_submit');
-      preventImplicitSubmit = prevent_implicit_submit;
-    }
+    preventImplicitSubmit = false;
   }
   if (revalidate === undefined) {
     /* other recognized values: 'oninput', 'onblur', 'onsubmit' and 'never' */
     revalidate = strict? 'onsubmit' : 'hybrid';
   }
   if (validEvent === undefined) {
-    if (valid_event === undefined) {
-      validEvent = ! strict;
-    } else {
-      w('valid_event');
-      validEvent = valid_event;
-    }
+    validEvent = ! strict;
   }
 
   const settings = { debug, strict, preventImplicitSubmit, revalidate,
@@ -126,15 +92,6 @@ hyperform.addValidator = (element, validator) => { CustomValidatorRegistry.set(e
 hyperform.setMessage = (element, validator, message) => { custom_messages.set(element, validator, message); return hyperform; };
 hyperform.addHook = (hook, action, position) => { add_hook(hook, action, position); return hyperform; };
 hyperform.removeHook = (hook, action) => { remove_hook(hook, action); return hyperform; };
-
-// TODO: Remove in next non-patch version
-hyperform.set_language = lang => { w('set_language'); set_language(lang); return hyperform; };
-hyperform.add_translation = (lang, catalog) => { w('add_translation'); add_translation(lang, catalog); return hyperform; };
-hyperform.set_renderer = (renderer, action) => { w('set_renderer'); Renderer.set(renderer, action); return hyperform; };
-hyperform.add_validator = (element, validator) => { w('add_validator'); CustomValidatorRegistry.set(element, validator); return hyperform; };
-hyperform.set_message = (element, validator, message) => { w('set_message'); custom_messages.set(element, validator, message); return hyperform; };
-hyperform.add_hook = (hook, action, position) => { w('add_hook'); add_hook(hook, action, position); return hyperform; };
-hyperform.remove_hook = (hook, action) => { w('remove_hook'); remove_hook(hook, action); return hyperform; };
 
 if (document.currentScript && document.currentScript.hasAttribute('data-hf-autoload')) {
   hyperform(window);